move create_synthetic_czi
