+++ webknossos/tests/utils.py
@@ -32,40 +32,3 @@ def TestTemporaryDirectoryNonLocal() -> Generator[UPath, None, None]:
     temp_dir = UPath(f"memory:///{random_prefix}")
     temp_dir.mkdir(parents=True, exist_ok=True)
     yield temp_dir
-
-
-def create_synthetic_czi(
-    path: UPath,
-    *,
-    num_timepoints: int = 1,
-    num_czi_channels: int = 1,
-    z: int = 1,
-    y: int = 8,
-    x: int = 10,
-    dtype: DTypeLike = np.uint16,
-) -> np.ndarray:
-    """Writes a real .czi via pylibCZIrw and returns the data as
-    (t, czi_channel, z, y, x), so tests can compare against it directly.
-
-    Unlike the .ims helper this produces a genuine file rather than a stand-in,
-    so no monkeypatching is needed. Voxel values are unique per position and
-    per (t, czi_channel), which is what makes a mixed-up axis or a wrongly
-    offset roi visible rather than merely plausible.
-    """
-    from pylibCZIrw import czi as pyczi
-
-    data = np.arange(num_timepoints * num_czi_channels * z * y * x).reshape(
-        num_timepoints, num_czi_channels, z, y, x
-    )
-    data = (data % np.iinfo(np.uint16).max).astype(dtype)
-
-    with pyczi.create_czi(str(path)) as writer:
-        for t in range(num_timepoints):
-            for c in range(num_czi_channels):
-                for k in range(z):
-                    writer.write(
-                        # pylibCZIrw writes planes as (y, x, components)
-                        data[t, c, k][:, :, np.newaxis],
-                        plane={"T": t, "C": c, "Z": k},
-                    )
-    return data
